Implement proper printing logic
The previous print function was just a hack that used runtime type tags
to print the values correctly. Replace it with a correct implementation
that actually usually the `to_string` method provided through the
`printable` interface.

Issues: Some polymorphic methods need to generate more than one
implementation, 2 tests are still failing because of that.