Add a `naming` pass to and carry symbol table
Previously, the symbol table was constructed during parsing, and it
was used during the `naming` phase (that happened inline during
parsing), the symbol table was discarded, and a new table was
constructed for type checking.

This diff refactors the symbol table to contain both type and AST
information, and keeps the symbol table around, filling the variable
information in the naming pass, and the type information during type
checking.