Fix type variable collisions - take #2
When generating types from AST nodes, eagerly replace every type
variable with T0...TN, and put the alias on scope, so that the original
type variable (e.g. `u`) will map to the unique name (e.g. `T1`).

We carry on with generating types from AST nodes, and the next time we
find a type node, we check whether that type has been replaced, by
checking if there's an alias in scope, if so, we create a type with the
unique name instead of the one in the AST.

This diff also takes care of printing the types as the user defined,
instead of printing the virtual unique type, e.g. `T1`, by keeping a
reverse table of all the aliases created, and looking it up when
printing a type.