Fix generic resolution
Now that all type variables are unique, it's safe to update the original
entry in the scope, not redefine it in the current scope.

This fixes the cases when you have a function `f` that takes as argument
another function `g`: we need generic resolution of `g` to happen in a
separate scope, so it's variables don't leak to `f`, but we also want
any variable from `f` used in `t` to be properly updated.

Example:
fn f<t>(g: (int) -> t) -> t {
  g(42)
}

fn id<u>(x: u) -> t { u }

f(id)

When checking the call `f(id)` we need to check that the type of `id`
is compatible with the type of `g`, and for that we need to take into
account `u`.
We don't want the resolution of `u` to leak to the type checking of the
function call, but we want to take into account that, since `u` is going
to be an `int`, and the return type of `id` is `u`, `id` will return `int`,
which in turn means that the `t` in the return type of `g` has type `int`,
and that the return type of `f`, which was `t` is also going to be `int`.