Rename `Naming` module to `Reassoc` as it only deals with associativity
This module has only ever dealt with associativity and precedence of
operators, but the idea was that it would also deal with renaming at
some point, hence the name. As renaming got implemented as a separate
pass, it seems reasonable to rename this pass according to the only
thing it actually does: fix associativity.

Closes #42