Replace explicit scope management with context markers
Instead of passing multiple typing contexts around to control what's
visible inside and outside the current scope, it switches to using
context markers to indicate what parts of the context are temporary.
The API is that we start a marker, run the code that adds the temporary
bindings to the context (e.g. add the bindings for parameters of
a function), we mark the end of the marker, use the current context
(e.g. to check the body of the function) and the clear the marker, i.e.
we don't want the parameters of the function to be visible outside of
its definition.